STASHDEV-9827 Fix LOCAL mode hibernate caching.  * Rework `AbstractAccessDelegate` to factor out only the common    methods between the various `Delegate` classes.  * Truly implement READ_COMMITTED transactional semantics (i.e.,    handle `insert`, `update`, and `remove` as no-ops, and only    apply modifications to the cache on `afterInsert`,    `afterUpdate` and `unlockItem`) in all asynchronous cache    access strategies.  * For completeness, implement `NonStrictReadWrite` and `ReadOnly`    AccessDelegates roughly the same as their counterparts in    EHCache.  * In `LocalRegionCache`, improve handling of invalidations in the    following cases:     - post-commit deletion (in `unlockItem`),     - Collections that have a `VersionComparator` but null version,     - bulk / multi-table updates (in `unlockRegion`).  * Nodes do not have to listen to their own events published on the    TopicService.
