+++ hazelcast-hibernate/hazelcast-hibernate4/src/main/java/com/hazelcast/hibernate/access/AccessDelegate.java
@@ -43,7 +43,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param txTimestamp a timestamp prior to the transaction start time
      * @return the cached object or <tt>null</tt>
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     Object get(Object key, long txTimestamp) throws CacheException;
 
@@ -57,7 +57,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param version The item's version value
      * @return Were the contents of the cache actual changed by this operation?
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean insert(Object key, Object value, Object version) throws CacheException;
 
@@ -71,7 +71,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param version The item's version value
      * @return Were the contents of the cache actual changed by this operation?
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean afterInsert(Object key, Object value, Object version) throws CacheException;
 
@@ -86,7 +86,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param previousVersion The item's previous version value
      * @return Were the contents of the cache actual changed by this operation?
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean update(Object key, Object value, Object currentVersion, Object previousVersion) throws CacheException;
 
@@ -102,7 +102,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param lock            The lock previously obtained from {@link #lockItem}
      * @return Were the contents of the cache actual changed by this operation?
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean afterUpdate(Object key, Object value, Object currentVersion, Object previousVersion, SoftLock lock)
             throws CacheException;
@@ -116,7 +116,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param version     the item version number
      * @return <tt>true</tt> if the object was successfully cached
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean putFromLoad(Object key, Object value, long txTimestamp, Object version) throws CacheException;
 
@@ -131,7 +131,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param minimalPutOverride Explicit minimalPut flag
      * @return <tt>true</tt> if the object was successfully cached
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     boolean putFromLoad(Object key, Object value, long txTimestamp, Object version, boolean minimalPutOverride)
             throws CacheException;
@@ -142,7 +142,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      *
      * @param key The key of the item to remove
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void remove(Object key) throws CacheException;
 
@@ -150,7 +150,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * Called to evict data from the entire region
      *
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void removeAll() throws CacheException;
 
@@ -160,7 +160,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      *
      * @param key The key of the item to remove
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void evict(Object key) throws CacheException;
 
@@ -169,7 +169,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * isolation.
      *
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void evictAll() throws CacheException;
 
@@ -185,7 +185,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param version The item's current version value
      * @return A representation of our lock on the item; or null.
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     SoftLock lockItem(Object key, Object version) throws CacheException;
 
@@ -194,7 +194,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      *
      * @return A representation of our lock on the item; or null.
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     SoftLock lockRegion() throws CacheException;
 
@@ -206,7 +206,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      * @param key  The item key
      * @param lock The lock previously obtained from {@link #lockItem}
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void unlockItem(Object key, SoftLock lock) throws CacheException;
 
@@ -216,7 +216,7 @@ public interface AccessDelegate<T extends HazelcastRegion> {
      *
      * @param lock The lock previously obtained from {@link #lockRegion}
      * @throws org.hibernate.cache.CacheException
-     *          Propogated from underlying {@link org.hibernate.cache.spi.Region}
+     *          Propagated from underlying {@link org.hibernate.cache.spi.Region}
      */
     void unlockRegion(SoftLock lock) throws CacheException;
 }
