+++ hazelcast-hibernate/hazelcast-hibernate4/src/main/java/com/hazelcast/hibernate/local/LocalRegionCache.java
@@ -24,20 +24,16 @@ import com.hazelcast.core.MessageListener;
 import com.hazelcast.hibernate.CacheEnvironment;
 import com.hazelcast.hibernate.HazelcastTimestamper;
 import com.hazelcast.hibernate.RegionCache;
-import com.hazelcast.hibernate.serialization.ExpiryMarker;
 import com.hazelcast.hibernate.serialization.Expirable;
+import com.hazelcast.hibernate.serialization.ExpiryMarker;
 import com.hazelcast.hibernate.serialization.Value;
+import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
 import com.hazelcast.util.Clock;
 import org.hibernate.cache.spi.CacheDataDescription;
 import org.hibernate.cache.spi.access.SoftLock;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -60,6 +56,8 @@ public class LocalRegionCache implements RegionCache {
     protected final Comparator versionComparator;
     protected final AtomicLong markerIdCounter;
     protected MapConfig config;
+
+    private final ILogger log = Logger.getLogger(LocalRegionCache.class);
 
     /**
      * @param name              the name for this region cache, which is also used to retrieve configuration/topic
@@ -91,7 +89,7 @@ public class LocalRegionCache implements RegionCache {
         try {
             config = hazelcastInstance != null ? hazelcastInstance.getConfig().findMapConfig(name) : null;
         } catch (UnsupportedOperationException e) {
-            Logger.getLogger(LocalRegionCache.class).finest(e);
+            log.finest(e);
         }
         versionComparator = metadata != null && metadata.isVersioned() ? metadata.getVersionComparator() : null;
         cache = new ConcurrentHashMap<Object, Expirable>();
@@ -127,11 +125,11 @@ public class LocalRegionCache implements RegionCache {
                 }
             } else if (previous.isReplaceableBy(txTimestamp, version, versionComparator)) {
                 if (cache.replace(key, previous, newValue)) {
-        return true;
-    }
+                    return true;
+                }
             } else {
-            return false;
-        }
+                return false;
+            }
         }
     }
 
@@ -195,18 +193,28 @@ public class LocalRegionCache implements RegionCache {
     protected MessageListener<Object> createMessageListener() {
         return new MessageListener<Object>() {
             public void onMessage(final Message<Object> message) {
+                if (message.getPublishingMember().localMember()) {
+                    return;
+                }
                 final Invalidation invalidation = (Invalidation) message.getMessageObject();
-                if (versionComparator != null) {
-                    final Expirable value = cache.get(invalidation.getKey());
-                    if (value != null) {
-                        Object currentVersion = value.getVersion();
-                        Object newVersion = invalidation.getVersion();
-                        if (versionComparator.compare(newVersion, currentVersion) > 0) {
-                            cache.remove(invalidation.getKey(), value);
+                Object key = invalidation.getKey();
+                if (key != null) {
+                    if (versionComparator != null) {
+                        final Expirable value = cache.get(key);
+                        if (value != null) {
+                            Object newVersion = invalidation.getVersion();
+                            if (newVersion != null) {
+                                Object currentVersion = value.getVersion();
+                                if (versionComparator.compare(newVersion, currentVersion) > 0) {
+                                    cache.remove(key, value);
+                                    return;
+                                }
+                            }
                         }
                     }
+                    cache.remove(key);
                 } else {
-                    cache.remove(invalidation.getKey());
+                    cache.clear();
                 }
             }
         };
@@ -214,11 +222,8 @@ public class LocalRegionCache implements RegionCache {
 
     public boolean remove(final Object key) {
         final Expirable value = cache.remove(key);
-        if (value != null) {
-            maybeNotifyTopic(key, null, value.getVersion());
-            return true;
-        }
-        return false;
+        maybeNotifyTopic(key, null, (value != null) ? value.getVersion() : null);
+        return (value != null);
     }
 
     public SoftLock tryLock(final Object key, final Object version) {
@@ -262,6 +267,7 @@ public class LocalRegionCache implements RegionCache {
                 break;
             }
         }
+        maybeNotifyTopic(key, null, null);
     }
 
     public boolean contains(final Object key) {
@@ -270,6 +276,7 @@ public class LocalRegionCache implements RegionCache {
 
     public void clear() {
         cache.clear();
+        maybeNotifyTopic(null, null, null);
     }
 
     public long size() {
